Added new env functions
