+++ log.log
@@ -1,0 +1,16 @@
+total 2776
+drwxr-xr-x   15 jules  staff      480 Oct 28 14:30 .
+drwxr-xr-x   15 jules  staff      480 Oct 28 14:12 ..
+drwxr-xr-x   16 jules  staff      512 Oct 28 14:26 .git
+drwxr-xr-x    3 jules  staff       96 Oct 24 13:29 .github
+-rw-r--r--    1 jules  staff       42 Oct 24 15:00 .gitignore
+-rw-r--r--    1 jules  staff       87 Oct 24 13:57 .gitmodules
+-rw-r--r--    1 jules  staff     2814 Oct 28 14:29 Makefile
+-rw-r--r--    1 jules  staff       11 Oct 24 13:29 README
+-rwxr-xr-x    1 jules  staff    12916 Oct 28 11:17 a.out
+drwxr-xr-x    9 jules  staff      288 Oct 28 14:26 includes
+drwxr-xr-x  127 jules  staff     4064 Oct 28 14:30 libft
+-rw-r--r--    1 jules  staff        0 Oct 28 14:30 log.log
+-rwxr-xr-x    1 jules  staff    85588 Oct 28 14:30 minishell
+-rw-r--r--@   1 jules  staff  1298894 Oct 24 13:29 minishell.en.subject.pdf
+drwxr-xr-x   41 jules  staff     1312 Oct 28 14:29 srcs
